requirements.txt
pandas==1.5.3
setuptools==60.2.0
numpy==1.23.3
pyxlsb==1.0.10
openpyxl==3.1.0
xlrd==2.0.1
notifiers==1.3.3
requests~=2.31.0
clickhouse-connect==0.5.14
python-dotenv==1.0.0